requirements.txt
# X/Twitter Scraping
snscrape

# AI Content Generation
anthropic

# Utilities
python-dotenv
requests

# Date handling
python-dateutil
